debuging

+++ bvar/test.py
@@ -31,13 +31,13 @@ X_star = pd.read_csv('Xstar.csv', delimiter=',',
 W = pd.read_csv('weight.csv', delimiter=',', 
                 dtype=float, header=0, usecols=range(2,41))
 
-data = np.array(Y)
-z = np.array(X_star)
-w = np.array(W)
+data = np.array(Y, dtype=float)
+z = np.array(X_star, dtype=float)
+w = np.array(W, dtype=float)
 
 os.chdir(MODULE_PATH)
 from model import FactorAugumentedVARX
 favarx = FactorAugumentedVARX(n_iter=100, n_save=50, lag=1, 
                               var_lag=1, n_factor=3, horizon=10,
-                              smoother_option='CarterKohn', 
+                              smoother_option='CarterKohn',
                               is_standardize=False).estimate(data, z, w)
